+++ src/Controller/MenuController.php
@@ -10,21 +10,39 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Bundle\SecurityBundle\Security;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 
 #[Route('/menu')]
 class MenuController extends AbstractController
 {
     #[Route('/', name: 'app_menu_index', methods: ['GET'])]
-    public function index(MenuRepository $menuRepository): Response
+    public function index(MenuRepository $menuRepository, Security $security): Response
     {
+        $user = $security->getUser();
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return new RedirectResponse('/');
+        }
         return $this->render('menu/index.html.twig', [
             'menus' => $menuRepository->findAll(),
         ]);
     }
+    #[Route('/show', name: 'app_menu_student', methods: ['GET'])]
+    public function showMenu(MenuRepository $menuRepository, Security $security): Response
+    {
+        $menus = $menuRepository->findFutureMenus();
+        return $this->render('menu/indexstudent.html.twig', [
+            'menus' => $menus,
+        ]);
+    }
+
 
     #[Route('/new', name: 'app_menu_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, Security $security): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return new RedirectResponse('/');
+        }
         $menu = new Menu();
         $form = $this->createForm(MenuType::class, $menu);
         $form->handleRequest($request);
@@ -43,16 +61,24 @@ class MenuController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_menu_show', methods: ['GET'])]
-    public function show(Menu $menu): Response
+    public function show(Menu $menu, Security $security): Response
     {
+        $user = $security->getUser();
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return new RedirectResponse('/');
+        }
         return $this->render('menu/show.html.twig', [
             'menu' => $menu,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_menu_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Menu $menu, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, Menu $menu, EntityManagerInterface $entityManager, Security $security): Response
     {
+        $user = $security->getUser();
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return new RedirectResponse('/');
+        }
         $form = $this->createForm(MenuType::class, $menu);
         $form->handleRequest($request);
 
@@ -71,6 +97,9 @@ class MenuController extends AbstractController
     #[Route('/{id}', name: 'app_menu_delete', methods: ['POST'])]
     public function delete(Request $request, Menu $menu, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return new RedirectResponse('/');
+        }
         if ($this->isCsrfTokenValid('delete'.$menu->getId(), $request->getPayload()->get('_token'))) {
             $entityManager->remove($menu);
             $entityManager->flush();
